Deleted raw file plot example: it has to be rewritten completely.
SOLVED(?) the network reconnect bug by resetting the asio accetor 
instead of deleting it. I have no idea if this is dependent on the boost 
version (here 1.68)

solved a few bugs releted to the dataset length using a chirped 
waveform.

solved a potential compatibility issue with numpy throwing a 
VisibleDeprecationWarning on VNA analyse calls.